app/requirements.txt
click==8.1.3; python_version >= "3.7"
colorama==0.4.5; python_version >= "3.7" and python_full_version < "3.0.0" and platform_system == "Windows" or platform_system == "Windows" and python_version >= "3.7" and python_full_version >= "3.5.0"
flask==2.2.2; python_version >= "3.7"
importlib-metadata==4.12.0; python_version < "3.8" and python_version >= "3.7"
itsdangerous==2.1.2; python_version >= "3.7"
jinja2==3.1.2; python_version >= "3.7"
markupsafe==2.1.1; python_version >= "3.7"
typing-extensions==4.3.0; python_version < "3.8" and python_version >= "3.7"
werkzeug==2.2.2; python_version >= "3.7"
zipp==3.8.1; python_version < "3.8" and python_version >= "3.7"
mysql-connector-python==8.0.30
